Setup basic frontend
