--- components.d.ts
@@ -1,13 +1,0 @@
-// generated by unplugin-vue-components
-// We suggest you to commit this file into source control
-// Read more: https://github.com/vuejs/core/pull/3399
-import '@vue/runtime-core'
-
-export {}
-
-declare module '@vue/runtime-core' {
-  export interface GlobalComponents {
-    FormContainer: typeof import('./resources/js/components/form/Container.vue')['default']
-    FormRow: typeof import('./resources/js/components/form/Row.vue')['default']
-  }
-}
