+++ package.json
@@ -1,5 +1,11 @@
 {
-  "private": true,
+  "name": "groups",
+  "description": "Groups",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/fsr5-fhaachen/groups/"
+  },
   "scripts": {
     "dev": "vite",
     "build": "vite build"
@@ -12,6 +18,7 @@
     "lodash": "^4.17.21",
     "postcss": "^8.4.16",
     "tailwindcss": "^3.1.8",
+    "unplugin-auto-import": "^0.11.2",
     "unplugin-vue-components": "^0.22.4",
     "vite": "^3.0.9",
     "vue-loader": "^17.0.0",
@@ -29,9 +36,9 @@
     "@inertiajs/inertia": "^0.11.0",
     "@inertiajs/inertia-vue3": "^0.6.0",
     "@inertiajs/progress": "^0.2.7",
+    "@tailwindcss/forms": "^0.5.2",
     "bootstrap": "^5.2.0",
     "bootstrap-vue": "^2.22.0",
-    "@tailwindcss/forms": "^0.5.2",
     "vue": "^3.2.37"
   }
 }
