+++ resources/js/composables/useSelectFormCourseOptions.ts
@@ -1,14 +1,12 @@
-import { FormKitOptionsList } from '@formkit/vue';
-
 export default (courses: any[]) => {
-  const options: FormKitOptionsList = [];
+  const options: Form.SelectOption[] = [];
 
   courses.forEach((course: any) => {
     options.push({
       value: course.id,
       label: course.name,
     });
-  }
+  });
 
   return options;
-}
+};
