+++ resources/js/layouts/DefaultLayout.vue
@@ -1,0 +1,7 @@
+<template>
+  <div class="bg-red-500">
+    <slot />
+  </div>
+</template>
+
+<script setup lang="ts"></script>
