+++ resources/js/types/auto-imports.d.ts
@@ -1,0 +1,5 @@
+// Generated by 'unplugin-auto-import'
+export {}
+declare global {
+  const useSelectFormCourseOptions: typeof import('../composables/useSelectFormCourseOptions')['default']
+}
