+++ resources/js/types/components.d.ts
@@ -1,0 +1,17 @@
+// generated by unplugin-vue-components
+// We suggest you to commit this file into source control
+// Read more: https://github.com/vuejs/core/pull/3399
+import '@vue/runtime-core'
+
+export {}
+
+declare module '@vue/runtime-core' {
+  export interface GlobalComponents {
+    Card: typeof import('./../layouts/Card.vue')['default']
+    CardLayout: typeof import('./../layouts/CardLayout.vue')['default']
+    Default: typeof import('./../layouts/Default.vue')['default']
+    DefaultLayout: typeof import('./../layouts/DefaultLayout.vue')['default']
+    FormContainer: typeof import('./../components/form/Container.vue')['default']
+    FormRow: typeof import('./../components/form/Row.vue')['default']
+  }
+}
