+++ resources/js/types/form.d.ts
@@ -1,0 +1,6 @@
+declare namespace Form {
+  export interface SelectOption {
+    value: string;
+    label: string;
+  }
+}
