+++ resources/js/types/global.d.ts
@@ -1,0 +1,5 @@
+export {};
+declare global {
+  const __PACKAGE_NAME__: string;
+  const __PACKAGE_REPOSITORY_URL__: string;
+}
